% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-atlas.R
\name{add_atlas}
\alias{add_atlas}
\title{Add atlas}
\usage{
add_atlas(img, png, rsx_trans, ..., props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{png}{A raw vector of PNG image to be used as a sprite.}

\item{rsx_trans}{A double matrix with 6 columns where each row represents an RSX transform.
Each column of the matrix corresponds to:
\itemize{
\item scale
\item angle of rotation (in radians)
\item amount of translation in the X-axis direction
\item amount of translation in the Y-axis direction
\item X coordinate of the anchor point
\item Y coordinate of the anchor point
}}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Draws the same number of sprites as the number of rows in \code{rsx_trans}.
}
