% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skiagd-package.R
\name{add_circle}
\alias{add_circle}
\title{Add a circle}
\usage{
add_circle(img, x, y, r, props = paint(), .size = dev_size())
}
\arguments{
\item{img}{a raw vector.}

\item{x}{a numeric.}

\item{y}{a numeric.}

\item{r}{a numeric.}

\item{props}{a paint prop.}

\item{.size}{an integer vector.}
}
\value{
a raw vector.
}
\description{
Add a circle
}
