% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-rect.R
\name{add_diff_rect}
\alias{add_diff_rect}
\title{Add difference rectangles}
\usage{
add_diff_rect(
  img,
  outer,
  outer_radii,
  inner,
  inner_radii,
  ...,
  props = paint()
)
}
\arguments{
\item{img}{A raw vector of picture.}

\item{outer}{A double matrix where each row is an outer rectangle
XYWH (left, top, right, bottom).}

\item{outer_radii}{A double matrix where each row is a pair of axis lengths
on X-axis and Y-axis of outer oval describing rounded corners.}

\item{inner}{A double matrix where each row is an inner rectangle
XYWH (left, top, right, bottom).}

\item{inner_radii}{A double matrix where each row is a pair of axis lengths
on X-axis and Y-axis of inner oval describing rounded corners.}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not take care of \code{...}, they are simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add difference rectangles
}
