% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{add_line}
\alias{add_line}
\title{Add lines}
\usage{
add_line(img, from, to, props = paint(), .size = dev_size())
}
\arguments{
\item{img}{A raw vector.}

\item{from}{A double matrix where each row is a start point.}

\item{to}{A double matrix where each row is an end point.}

\item{props}{A paint properties out of \code{\link[=paint]{paint()}}.}

\item{.size}{An integer vector.}
}
\value{
a raw vector.
}
\description{
Add lines
}
