% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{add_rect}
\alias{add_rect}
\title{Add rectangles}
\usage{
add_rect(img, rect, props = paint(), .size = dev_size())
}
\arguments{
\item{img}{A raw vector.}

\item{rect}{A double matrix where each row is a rectangle corner.}

\item{props}{A paint properties out of \code{\link[=paint]{paint()}}.}

\item{.size}{An integer vector.}
}
\value{
A raw vector.
}
\description{
Add rectangles
}
