% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-textblob.R
\name{add_text}
\alias{add_text}
\alias{text_info}
\title{Add text}
\usage{
add_text(img, text, rsx_trans, ..., props = paint())

text_info(text, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{text}{Characters to be drawn.}

\item{rsx_trans}{A double matrix where each row represents an RSX transform.
Each column of the matrix corresponds to the scale, the angle of rotation,
the amount of translation
in the X-axis direction and in the Y-axis direction,
and the X and Y coordinates of the pivot point.}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
For \code{add_text()}, a raw vector of picture.
}
\description{
Draws text as textblobs.

The return value is often a large object
because the specified font is embedded in the returned picture.
Note that you should almost always \code{\link[=freeze]{freeze()}} the picture after drawing text.
}
\details{
Since textblobs do not have font fallback mechanism,
characters out of the specified font are not drawn correctly.
}
