% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native-raster.R
\name{as_nativeraster}
\alias{as_nativeraster}
\title{Convert picture into native raster}
\usage{
as_nativeraster(img, ..., props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{...}{For some drawing functions, you can specify
\code{sigma}, \code{width} and \code{color} as named arguments.
\itemize{
\item \code{sigma} must be a numeric vector of blur sigmas for each shape.
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A native raster.
}
\description{
Convert picture into native raster
}
\note{
This function returns integer matrices of \code{nativeRaster} class,
but they do not strictly conform to R's native raster specification.

Since skiagd internally uses premultiplied alpha,
these native raster objects may look different
than the expected PNG image if the alpha channel is not fully opaque.
}
