% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{as_png}
\alias{as_png}
\title{Convert picture into PNG image}
\usage{
as_png(img, ..., props = paint())
}
\arguments{
\item{img}{A raw vector of a serialized Skia picture.}

\item{...}{For some drawing functions, you can specify
\code{sigma}, \code{width} and \code{color} as named arguments.
\itemize{
\item \code{sigma} must be a numeric vector of blur sigmas for each shape.
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix with 4 rows (RGBA) and N columns (shapes).
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{props}{A list of painting attributes created by \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of a PNG image.
}
\description{
Renders a serialized picture to a PNG image.
}
\examples{
\dontrun{
png <-
 canvas("navy") |>
 as_png()

# Write the PNG image to a file
writeBin(png, "navy.png")
}
}
