% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{as_recordedplot}
\alias{as_recordedplot}
\title{Convert picture into recorded plot}
\usage{
as_recordedplot(img, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.
For drawing functions, this can be a list of several painting attributes;
if it is a nested list, the drawing operation is performed
sequentially using each one.}
}
\value{
A \code{recordedplot} object is invisibly returned.
See \code{\link[grDevices:recordplot]{grDevices::recordPlot()}} for details.
}
\description{
This is mainly for testing purposes.
}
