% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{BlendMode}
\alias{BlendMode}
\title{BlendMode (0-28)}
\format{
An object of class \code{BlendMode__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 29.
}
\usage{
BlendMode
}
\description{
BlendMode (0-28)
}
\seealso{
\href{https://rust-skia.github.io/doc/skia_safe/enum.BlendMode.html}{BlendMode in skia_safe - Rust}
}
\keyword{datasets}
