% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skiagd-package.R
\name{canvas}
\alias{canvas}
\title{Create a canvas}
\usage{
canvas(fill = paint()[["fill"]], .size = dev_size())
}
\arguments{
\item{fill}{a color name.}
}
\value{
a raw vector.
}
\description{
Creates a new canvas filled with specified color.
}
