% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{Cap}
\alias{Cap}
\title{Cap (0-2)}
\format{
An object of class \code{Cap__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 3.
}
\usage{
Cap
}
\description{
Cap (0-2)
}
\seealso{
\href{https://rust-skia.github.io/doc/skia_safe/paint/enum.Cap.html}{Cap in skia_safe::paint - Rust}
}
\keyword{datasets}
