% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{create_mapping}
\alias{create_mapping}
\title{Create transformation matrix from points pairs}
\usage{
create_mapping(src, dst)
}
\arguments{
\item{src}{A double matrix where each row is a point.}

\item{dst}{A double matrix where each row is a point.}
}
\value{
A numeric matrix of size 3x3.
}
\description{
Creates a transformation matrix from two sets of points.
If mapping is not possible, throws an error.
}
