% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{dev_size}
\alias{dev_size}
\title{Get the size of the current graphics device}
\usage{
dev_size(units = "px")
}
\arguments{
\item{units}{\code{units} for \code{\link[grDevices:dev.size]{grDevices::dev.size()}}.}
}
\value{
An integer vector of length 2 (width, height).
}
\description{
Just returns the size of the current graphics device as integers.
}
