% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-lock-environments.R
\name{enable_autocomplete}
\alias{enable_autocomplete}
\title{Enable autocomplete for painting attributes}
\usage{
enable_autocomplete(env = parent.frame())
}
\arguments{
\item{env}{Environment to assign objects to.}
}
\value{
Called for its side-effect.
}
\description{
Enables autocomplete for painting attributes.
To to this, this function simply assigns them to the current environment.
}
