% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{\%timer\%}
\alias{\%timer\%}
\title{Log execution time of an expression}
\usage{
lhs \%timer\% rhs
}
\arguments{
\item{lhs}{An expression.}

\item{rhs}{An expression.}
}
\value{
Values from evaluated \code{rhs} is returned invisibly.
}
\description{
An operator to evaluate an expression
while logging its execution time for debugging.
}
\details{
\code{lhs \%timer\% rhs} evaluates \code{rhs} inside a \link{system.time} call,
assigns the execution time to the variable \code{time},
and tries to evaluate \code{lhs} in an environment where \code{time} exists.
In doing so, failure to evaluate the left-hand side
does not result in an error.
}
\examples{
print(time) \%timer\% {
 rnorm(10) ^2
}
}
