% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{Join}
\alias{Join}
\title{Join (0-2)}
\format{
An object of class \code{Join__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 3.
}
\usage{
Join
}
\description{
Join (0-2)
}
\seealso{
\href{https://rust-skia.github.io/doc/skia_safe/paint/enum.Join.html}{Join in skia_safe::paint - Rust}
}
\keyword{datasets}
