% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skiagd-package.R
\name{list_font_families}
\alias{list_font_families}
\title{List available font families}
\usage{
list_font_families()
}
\value{
A tibble.
}
\description{
Returns font families available on the system.

Since skiagd can only access fonts installed on the system,
font families in the font registry or local fonts
registered by the \href{https://systemfonts.r-lib.org/}{systemfonts} package
cannot be specified as the \code{family} in \code{\link[=paint]{paint()}}.
}
