% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{op_count}
\alias{op_count}
\title{Returns the approximate number of operations in the picture}
\usage{
op_count(picture, nested)
}
\arguments{
\item{picture}{A raw vector of picture.}

\item{nested}{A logcial scalar; if \code{TRUE}, counts nested operations.}
}
\description{
Returns the approximate number of operations in the picture
}
\keyword{internal}
