% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skiagd-package.R
\name{paint}
\alias{paint}
\title{Paint props}
\usage{
paint(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> What these dots do.}
}
\value{
a list.
}
\description{
Paint props
}
