% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-utils.R
\name{path_transform}
\alias{path_transform}
\title{Transform SVG path}
\usage{
path_transform(path, transform)
}
\arguments{
\item{path}{Characters of SVG notations.}

\item{transform}{Numerics of length 9.}
}
\value{
A character vector of SVG notations.
}
\description{
Transform SVG path
}
\seealso{
Other path-utils: 
\code{\link{path_bounds}()},
\code{\link{path_interpolate}()}
}
\concept{path-utils}
