% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{PathEffect}
\alias{PathEffect}
\title{PathEffect}
\format{
An object of class \code{PathEffect__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 8.
}
\usage{
PathEffect
}
\description{
\code{PathEffect} is a struct that offers a reference to \code{skia_safe::PathEffect}.
You can apply a path effect to shapes via \code{\link[=paint]{paint()}}.
Currently only single \code{PathEffect} can be specified; multiple effects are not supported.
}
\details{
The following effects are available:
\itemize{
\item \code{no_effect()}: does not apply any path effect. This is the default effect for \code{paint()}.
\item \code{trim(start, end)}: trims the \code{start} and \code{end} of the path. \code{start} and \code{end} are in the range \verb{[0, 1]}.
\item \code{discrete(length, deviation, seed)}: applies discrete path effect.
\item \code{dash(intervals, phase)}: applies dash path effect.
\item \code{corner(radius)}: applies corner path effect.
\item \code{path_1d(path, advance, phase, style)}: applies 1D path effect. \code{style} can be \code{"translate"}, \code{"rotate"}, or \code{"morph"}.
\item \code{path_2d(path, mat)}: applies 2D path effect.
\item \code{line_2d(width, mat)}: applies 2D line path effect.
}
}
\seealso{
\href{https://shopify.github.io/react-native-skia/docs/path-effects/}{Path Effects | React Native Skia}
}
\keyword{datasets}
