% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{pictures}
\alias{pictures}
\alias{picture}
\title{Pictures}
\description{
In Skia, a picture is a prerecorded list of drawing operations on a canvas.
The drawing functions of skiagd
take it as their first argument, add new shapes onto it,
and return a serialized picture as a raw vector again.

A serialized picture is a binary format containing a single frame,
which can be saved to a \code{.skp} file using \code{\link[=writeBin]{writeBin()}},
and reused by any drawing functions
as long as it is compatible with the version of Skia
used to create it.

You can review contents of \code{.skp} files
with the \href{https://skia.org/docs/dev/tools/debugger/}{Skia debugger}
if they are compatible with the version.
}
\section{Notes}{

\itemize{
\item A serialized picture can hold only up to 100 drawing operations.
If you need more, you can save the capacity by using \code{\link[=freeze]{freeze()}} as needed.
}
}

\seealso{
\href{https://shopify.github.io/react-native-skia/docs/shapes/pictures}{Pictures | React Native Skia}
}
