% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{PointMode}
\alias{PointMode}
\title{PointMode (0-2)}
\format{
An object of class \code{PointMode__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 3.
}
\usage{
PointMode
}
\description{
This is for \code{\link[=add_point]{add_point()}} only.
Not used in other functions.
}
\seealso{
\href{https://rust-skia.github.io/doc/skia_safe/canvas/enum.PointMode.html}{PointMode in skia_safe::canvas - Rust}
}
\keyword{datasets}
