% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skiagd-package.R
\name{save_img}
\alias{save_img}
\title{Save image to file}
\usage{
save_img(img, filename, .size = dev_size())
}
\arguments{
\item{img}{a raw vector.}

\item{filename}{a character string.}

\item{.size}{an integer vector.}
}
\value{
\code{filename} is invisibly returned.
}
\description{
Save image to file
}
