% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{Shader}
\alias{Shader}
\title{Shader}
\format{
An object of class \code{Shader__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 10.
}
\usage{
Shader
}
\description{
Shader
}
\seealso{
\itemize{
\item \href{https://shopify.github.io/react-native-skia/docs/shaders/gradients}{Gradients | React Native Skia}
\item \href{https://shopify.github.io/react-native-skia/docs/shaders/perlin-noise}{Perlin Noise Shaders | React Native Skia}
\item \href{https://shopify.github.io/react-native-skia/docs/shaders/colors}{Blending and Colors | React Native Skia}
}
}
\keyword{datasets}
