% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{BlendMode}
\alias{BlendMode}
\title{BlendMode (0-28)}
\format{
An object of class \code{BlendMode__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 29.
}
\usage{
BlendMode
}
\description{
\code{BlendMode} determines how source and destination colors are combined.
}
\details{
The following blend modes are available in Skia:
\enumerate{
\item \code{Clear}
\item \code{Src}
\item \code{Dst}
\item \code{SrcOver}
\item \code{DstOver}
\item \code{SrcIn}
\item \code{DstIn}
\item \code{SrcOut}
\item \code{DstOut}
\item \code{SrcATop}
\item \code{DstATop}
\item \code{Xor}
\item \code{Plus}
\item \code{Modulate}
\item \code{Screen}
\item \code{Overlay}
\item \code{Darken}
\item \code{Lighten}
\item \code{ColorDodge}
\item \code{ColorBurn}
\item \code{HardLight}
\item \code{SoftLight}
\item \code{Difference}
\item \code{Exclusion}
\item \code{Multiply}
\item \code{Hue}
\item \code{Saturation}
\item \code{Color}
\item \code{Luminosity}
}
}
\seealso{
\href{https://rust-skia.github.io/doc/skia_safe/enum.BlendMode.html}{BlendMode in skia_safe - Rust}

Other paint-attributes: 
\code{\link{Cap}},
\code{\link{FillType}},
\code{\link{FontStyle}},
\code{\link{Join}},
\code{\link{PathEffect}},
\code{\link{PointMode}},
\code{\link{Shader}},
\code{\link{Style}}
}
\concept{paint-attributes}
\keyword{datasets}
