% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{BlurStyle}
\alias{BlurStyle}
\title{BlurStyle (0-3)}
\format{
An object of class \code{skiagd::BlurStyle__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 4.
}
\usage{
BlurStyle
}
\description{
\code{BlurStyle} controls how a blur mask filter is applied to the shape.
}
\details{
The following \code{BlurStyle} are available:
\itemize{
\item \code{Normal}: Normal blur.
\item \code{Solid}: Solid blur.
\item \code{Outer}: Outer blur.
\item \code{Inner}: Inner blur.
}
}
\seealso{
\href{https://rust-skia.github.io/doc/skia_safe/enum.BlurStyle.html}{BlurStyle in skia_safe - Rust}

Other paint-attributes: 
\code{\link{BlendMode}},
\code{\link{Cap}},
\code{\link{FillType}},
\code{\link{FontStyle}},
\code{\link{ImageFilter}},
\code{\link{Join}},
\code{\link{PathEffect}},
\code{\link{PointMode}},
\code{\link{Shader}},
\code{\link{Style}},
\code{\link{VertexMode}}
}
\concept{paint-attributes}
\keyword{datasets}
