% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{FillType}
\alias{FillType}
\title{FillType (0-3)}
\format{
An object of class \code{FillType__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 4.
}
\usage{
FillType
}
\description{
\code{FillType} determines how paths are drawn.
This is for \code{\link[=add_path]{add_path()}} only. Not used in other functions.
}
\details{
The following \code{FillType} are available:
\itemize{
\item \code{Winding}
\item \code{EvenOdd}
\item \code{InverseWinding}
\item \code{InverseEvenOdd}
}
}
\seealso{
\href{https://rust-skia.github.io/doc/skia_safe/path/enum.FillType.html}{FillType in skia_safe::path - Rust}

Other paint-attributes: 
\code{\link{BlendMode}},
\code{\link{Cap}},
\code{\link{FontStyle}},
\code{\link{ImageFilter}},
\code{\link{Join}},
\code{\link{PathEffect}},
\code{\link{PointMode}},
\code{\link{Shader}},
\code{\link{Style}},
\code{\link{VertexMode}}
}
\concept{paint-attributes}
\keyword{datasets}
