% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{Join}
\alias{Join}
\title{Join (0-2)}
\format{
An object of class \code{Join__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 3.
}
\usage{
Join
}
\description{
\code{Join} determines the stroke join (the geometry drawn at the corners of strokes) for shapes.
}
\details{
The following joins are available:
\itemize{
\item \code{Miter}: Miter join.
\item \code{Round}: Round join.
\item \code{Bevel}: Bevel join.
}
}
\seealso{
\href{https://rust-skia.github.io/doc/skia_safe/paint/enum.Join.html}{Join in skia_safe::paint - Rust}

Other paint-attributes: 
\code{\link{BlendMode}},
\code{\link{Cap}},
\code{\link{FillType}},
\code{\link{PointMode}},
\code{\link{Style}}
}
\concept{paint-attributes}
\keyword{datasets}
