% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{Style}
\alias{Style}
\title{Style (0-2)}
\usage{
Style
}
\description{
\code{Style} determines the stroke style of shapes.
}
\details{
The following styles are available:
\itemize{
\item \code{StrokeAndFill}: Stroke and fill.
\item \code{Stroke}: Stroke only.
\item \code{Fill}: Fill only.
}
}
\seealso{
\href{https://rust-skia.github.io/doc/skia_safe/paint/enum.Style.html}{Style in skia_safe::paint - Rust}

Other paint-attributes:
\code{\link{BlendMode}},
\code{\link{BlurStyle}},
\code{\link{Cap}},
\code{\link{FillType}},
\code{\link{FontStyle}},
\code{\link{ImageFilter}},
\code{\link{Join}},
\code{\link{PathEffect}},
\code{\link{PointMode}},
\code{\link{Shader}},
\code{\link{VertexMode}}
}
\concept{paint-attributes}
