% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-filter.R
\name{ImageFilter}
\alias{ImageFilter}
\title{ImageFilter}
\arguments{
\item{source}{A \link{RuntimeEffect} object.}

\item{uniforms}{A named list of numerics to be assigned to uniforms in \code{source}.}
}
\value{
An \code{ImageFilter} object.
}
\description{
\code{ImageFilter} is a struct that offers a reference to \code{skia_safe::ImageFilter}.
You can apply an image filter to canvas via \code{\link[=paint]{paint()}}.
}
\details{
The following filters are available:
\itemize{
\item \code{no_filter()}: does not apply any image filter. This is the default image filter for \code{\link[=paint]{paint()}}.
\item \code{runtime_shader(source, uniforms)}: Creates an image filter from a \link{RuntimeEffect}.
}
}
\seealso{
\href{https://shopify.github.io/react-native-skia/docs/image-filters/overview/}{Image Filters | React Native Skia}

Other paint-attributes: 
\code{\link{BlendMode}},
\code{\link{BlurStyle}},
\code{\link{Cap}},
\code{\link{FillType}},
\code{\link{FontStyle}},
\code{\link{Join}},
\code{\link{PathEffect}},
\code{\link{PointMode}},
\code{\link{Shader}},
\code{\link{Style}},
\code{\link{VertexMode}}
}
\concept{paint-attributes}
