% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skiagd-package.R
\docType{package}
\name{skiagd-package}
\alias{skiagd}
\alias{skiagd-package}
\title{skiagd: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
