% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{param-img-and-props}
\alias{param-img-and-props}
\alias{param-rsx-trans}
\title{Params for drawing functions}
\arguments{
\item{img}{A raw vector of picture.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{rsx_trans}{A double matrix with 6 columns where each row represents an RSX transform.
Each column of the matrix corresponds to the scale, the angle of rotation (in radians),
the amount of translation in the X-axis direction and in the Y-axis direction,
and the X and Y coordinates of the transformation origin.}
}
\description{
Params for drawing functions

RSX transform
}
\keyword{internal}
