% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-effect.R
\name{PathEffect}
\alias{PathEffect}
\title{PathEffect}
\arguments{
\item{start}{A numeric scalar in the range \verb{[0, 1]}.}

\item{end}{A numeric scalar in the range \verb{[0, 1]}.}

\item{length}{A numeric scalar; length of the subsegments.}

\item{deviation}{A numeric scalar; limit of the movement of the endpoints.}

\item{seed}{An integer scalar; random seed.}

\item{intervals}{A numeric vector; even number of entries with even indices
specifying the length of the "on" intervals,
and the odd index specifying the length of "off".}

\item{phase}{A numeric scalar; offset into the intervals array (for \code{dash()}),
or distance (mod advance) along the path for its initial position (for \code{path_1d()}).}

\item{radius}{A numeric scalar; radius of the rounded corners.}

\item{path}{A string scalar of SVG notation to replicate.}

\item{advance}{A numeric scalar; space between instances of path.}

\item{style}{A string scalar; how to transform path at each point.
Can be \code{"translate"}, \code{"rotate"}, or \code{"morph"}.}

\item{transform}{Numerics of length 9; transformation matrix.}

\item{width}{A numeric scalar; width of the path to be stamped.}
}
\value{
A \code{PathEffect} object.
}
\description{
\code{PathEffect} is a struct that offers a reference to \code{skia_safe::PathEffect}.
You can apply a path effect to shapes via \code{\link[=paint]{paint()}}.
Currently only single \code{PathEffect} can be specified; multiple effects are not supported.
}
\details{
The following effects are available:
\itemize{
\item \code{no_effect()}: does not apply any path effect. This is the default effect for \code{\link[=paint]{paint()}}.
\item \code{trim(start, end)}: trims the \code{start} and \code{end} of the path. Note that you can't trim nothing at all, i.e., setting \code{start = 0} and \code{end = 1} arises an error.
\item \code{discrete(lentgh, deviation, seed)}: applies discrete path effect.
\item \code{dash(intervals, phase)}: applies dash path effect.
\item \code{corner(radius)}: applies corner path effect.
\item \code{path_1d(path, advance, phase, style)}: applies 1D path effect.
\item \code{path_2d(path, transform)}: applies 2D path effect.
\item \code{line_2d(width, transform)}: applies 2D line path effect.
}
}
\seealso{
\href{https://shopify.github.io/react-native-skia/docs/path-effects/}{Path Effects | React Native Skia}

Other paint-attributes: 
\code{\link{BlendMode}},
\code{\link{Cap}},
\code{\link{FillType}},
\code{\link{Join}},
\code{\link{PointMode}},
\code{\link{Shader}},
\code{\link{Style}}
}
\concept{paint-attributes}
