% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-utils.R
\name{path_transform}
\alias{path_transform}
\alias{path_interpolate}
\alias{path_bounds}
\title{Transform SVG path}
\usage{
path_transform(path, transform)

path_interpolate(t, first, second)

path_bounds(path)
}
\arguments{
\item{path}{Characters of SVG notations.}

\item{transform}{Numerics of length 9.}

\item{t}{A numeric vector of weights.
If it is inbetween or outside the range, the path is interpolated.}

\item{first}{A string scalar of SVG notation.}

\item{second}{A string scalar of SVG notation.}
}
\value{
For \code{path_transform()}: a character vector of SVG notations.

For \code{path_interpolate()}: a character vector of SVG notations interpolated.

For \code{path_bounds()}: a numeric matrix.
}
\description{
Interpolates between two SVG paths of equal length.
If the paths are not interpolatable, an error is thrown.

Returns minimum and maximum axes values of \code{path}.
}
