% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shader.R
\name{Shader}
\alias{Shader}
\title{Shader}
\description{
\code{Shader} is a struct that offers a reference to \code{skia_safe::Shader}.
You can apply a shader to paintings via \code{\link[=paint]{paint()}}.
}
\seealso{
\itemize{
\item \href{https://shopify.github.io/react-native-skia/docs/shaders/gradients}{Gradients | React Native Skia}
\item \href{https://shopify.github.io/react-native-skia/docs/shaders/perlin-noise}{Perlin Noise Shaders | React Native Skia}
\item \href{https://shopify.github.io/react-native-skia/docs/shaders/colors}{Blending and Colors | React Native Skia}
}
}
