% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-utils.R
\name{svg_bounds}
\alias{svg_bounds}
\title{Retrieve bounding boxes of SVG paths}
\usage{
svg_bounds(path)
}
\arguments{
\item{path}{A character vector of SVG path notations (the \code{d} attribute syntax).
Each element is treated as a separate path.}
}
\value{
A tibble containing columns \code{id}, \code{left}, \code{top}, \code{right} and \code{bottom} for each path.
\code{id} is a 1-based index corresponding to the input order of \code{path}.
}
\description{
Computes axis-aligned bounding boxes for one or more SVG paths.
}
\examples{
svg_bounds("M45 10 H55 V45 H90 V55 H55 V90 H45 V55 H10 V45 H45 Z")
}
\seealso{
Other path-utils: 
\code{\link{svg_interpolate}()},
\code{\link{svg_transform}()}
}
\concept{path-utils}
