% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-utils.R
\name{svg_transform}
\alias{svg_transform}
\title{Transform SVG paths}
\usage{
svg_transform(path, transform)
}
\arguments{
\item{path}{Characters of SVG notations.}

\item{transform}{Numerics of length 9.}
}
\value{
A character vector of SVG notations.
}
\description{
Transform SVG paths
}
\seealso{
Other path-utils: 
\code{\link{svg_bounds}()},
\code{\link{svg_interpolate}()}
}
\concept{path-utils}
