% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skiagd-package.R
\name{with_group}
\alias{with_group}
\title{Paint group}
\usage{
with_group(expr, ...)
}
\arguments{
\item{expr}{Expressions.}

\item{...}{Any other arguments are passed to \code{\link[=paint]{paint()}}.}
}
\description{
Evaluates \code{expr} with \link{paint(...)}.
}
