% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-arc.R
\name{add_arc}
\alias{add_arc}
\title{Add arcs}
\usage{
add_arc(
  img,
  ltrb,
  rsx_trans = matrix(c(1, 0, 0, 0, 0, 0), nrow(ltrb), 6, byrow = TRUE),
  angle = matrix(c(0, 360), nrow(ltrb), 2, byrow = TRUE),
  use_center = TRUE,
  ...,
  props = paint()
)
}
\arguments{
\item{img}{A raw vector of picture.}

\item{ltrb}{A double matrix where each row is a rectangle
LTRB (left, top, right, bottom).
Each rectangle bounds the area of oval containing arc to draw.}

\item{rsx_trans}{A double matrix with 6 columns where each row represents an RSX transform.
Each column of the matrix corresponds to:
\itemize{
\item scale
\item angle of rotation (in radians)
\item amount of translation in the X-axis direction
\item amount of translation in the Y-axis direction
\item X coordinate of the anchor point
\item Y coordinate of the anchor point
}}

\item{angle}{A double matrix where each row is a pair of sweeping angles (in degrees).}

\item{use_center}{Whether to draw a wedge that includes lines from oval center to arc end points.}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add arcs
}
