% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{add_circle}
\alias{add_circle}
\title{Add circles}
\usage{
add_circle(img, center, radius, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{center}{A double matrix where each row is circle center.}

\item{radius}{A double vector of circle radius.}

\item{props}{A paint properties out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector picture.
}
\description{
Add circles
}
