% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-rect.R
\name{add_diff_rect}
\alias{add_diff_rect}
\title{Add difference rectangles}
\usage{
add_diff_rect(img, outer, outer_radii, inner, inner_radii, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{outer}{A double matrix where each row is an outer rectangle
XYWH (left, top, right, bottom).}

\item{outer_radii}{A double matrix where each row is a pair of axis lengths
on X-axis and Y-axis of outer oval describing rounded corners.}

\item{inner}{A double matrix where each row is an inner rectangle
XYWH (left, top, right, bottom).}

\item{inner_radii}{A double matrix where each row is a pair of axis lengths
on X-axis and Y-axis of inner oval describing rounded corners.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add difference rectangles
}
