% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-line.R
\name{add_line}
\alias{add_line}
\title{Add lines}
\usage{
add_line(img, from, to, ..., props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{from}{A double matrix where each row is a start point.}

\item{to}{A double matrix where each row is an end point.}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add lines
}
