% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-path.R
\name{add_path}
\alias{add_path}
\title{Add paths}
\usage{
add_path(img, path, transform = diag(1, 3), ..., props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{path}{Characters of SVG notations
like \code{"M10 10 H 90 V 90 H 10 L 10 10"}.}

\item{transform}{Numerics of length 9
to apply affine transformations to the path themselves.
See \link{transform-matrix} for affine transformations.}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add paths
}
