% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{add_path}
\alias{add_path}
\title{Add paths}
\usage{
add_path(img, path, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{path}{A character vector of SVG strings.}

\item{props}{A paint properties out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add paths
}
