% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-path.R
\name{add_path}
\alias{add_path}
\title{Add paths}
\usage{
add_path(img, path, transform = c(1, 0, 0, 0, 1, 0, 0, 0, 1), props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{path}{A character vector of SVG notations
like \verb{M10 10 H 90 V 90 H 10 L 10 10}.}

\item{transform}{A numeric vector of length 9
to apply affine transformations to the path themselves.
See \link{transform-matrix} for affine transformations.}

\item{props}{A paint properties out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add paths
}
