% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{add_png}
\alias{add_png}
\title{Add PNG image to canvas}
\usage{
add_png(img, png, left = 0, top = 0, ..., props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{png}{A raw vector of PNG image.}

\item{left}{Left offset for drawing PNG image.}

\item{top}{Top offset for drawing PNG image.}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not take care of \code{...}, they are simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add PNG image to canvas
}
