% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{add_png}
\alias{add_png}
\title{Add PNG image to canvas}
\usage{
add_png(img, png, left = 0, top = 0, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{png}{A raw vector of PNG data.}

\item{left}{Left offset for drawing PNG image.}

\item{top}{Top offset for drawing PNG image.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add PNG image to canvas
}
