% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-point.R
\name{add_point}
\alias{add_point}
\title{Add points}
\usage{
add_point(img, point, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{point}{A double matrix where each row is a point.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.
For drawing functions, this can be a list of several painting attributes;
if it is a nested list, the drawing operation is performed
sequentially using each one.}
}
\value{
A raw vector of picture.
}
\description{
Add points
}
