% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{add_point}
\alias{add_point}
\title{Add points}
\usage{
add_point(
  img,
  point,
  mode = c("points", "lines", "polygon"),
  props = paint(),
  .size = dev_size()
)
}
\arguments{
\item{img}{A raw vector.}

\item{point}{A double matrix where each row is a point.}

\item{props}{A paint properties out of \code{\link[=paint]{paint()}}.}

\item{.size}{An integer vector.}
}
\value{
A raw vector.
}
\description{
Add points
}
