% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-rect.R
\name{add_rect}
\alias{add_rect}
\title{Add rectangles}
\usage{
add_rect(img, rect, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{rect}{A double matrix where each row is a rectangle
XYWH (left, top, right, bottom).}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.
For drawing functions, this can be a list of several painting attributes;
if it is a nested list, the drawing operation is performed
sequentially using each one.}
}
\value{
A raw vector of picture.
}
\description{
Add rectangles
}
