% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-rect.R
\name{add_rect}
\alias{add_rect}
\title{Add rectangles}
\usage{
add_rect(
  img,
  xywh,
  rsx_trans = matrix(c(1, 0, 0, 0, 0, 0), nrow(xywh), 6, byrow = TRUE),
  radii = matrix(0, nrow(xywh), 2),
  ...,
  props = paint()
)
}
\arguments{
\item{img}{A raw vector of picture.}

\item{xywh}{A double matrix where each row is a rectangle
XYWH (left, top, right, bottom).}

\item{rsx_trans}{A double matrix with 6 columns where each row represents an RSX transform.
Each column of the matrix corresponds to the scale, the angle of rotation (in radians),
the amount of translation in the X-axis direction and in the Y-axis direction,
and the X and Y coordinates of the transformation origin.}

\item{radii}{A double matrix where each row is a pair of axis lengths
on X-axis and Y-axis of oval describing rounded corners.}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add rectangles
}
