% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-rect.R
\name{add_rounded_rect}
\alias{add_rounded_rect}
\title{Add rounded rectangles}
\usage{
add_rounded_rect(img, rect, radii, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{rect}{A double matrix where each row is a rectangle
XYWH (left, top, right, bottom).}

\item{radii}{A double matrix where each row is a pair of axis lengths
on X-axis and Y-axis of oval describing rounded corners.}

\item{props}{A list of paint attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add rounded rectangles
}
