% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-textblob.R
\name{add_text}
\alias{add_text}
\alias{text_layout_horizontal}
\alias{text_layout_vertical}
\title{Add text}
\usage{
add_text(
  img,
  text,
  point = text_layout_horizontal(text, props),
  props = paint()
)

text_layout_horizontal(text, props = paint())

text_layout_vertical(text, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{text}{Characters to be drawn.}

\item{point}{A double matrix where each row is the point
at which each character in \code{text} is drawn.
For example, if \code{text} is a character vector of 5 and 3 length strings,
\code{point} must be a 8x2 matrix.}

\item{props}{A list of paint attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Draws text as textblobs.

The return value is often a large object
because the specified font is embedded in the returned picture.
Note that you should almost always \code{\link[=freeze]{freeze()}} the picture after drawing text.
}
\details{
You can use \code{text_layout_horizontal()} and \code{text_layout_vertical()}
to create a \code{point} matrix.
}
