% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-textblob.R
\name{add_text}
\alias{add_text}
\title{Add text}
\usage{
add_text(img, text, rsx_trans, ..., props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{text}{Characters to be drawn.}

\item{rsx_trans}{A double matrix with 6 columns where each row represents an RSX transform.
Each column of the matrix corresponds to the scale, the angle of rotation (in radians),
the amount of translation in the X-axis direction and in the Y-axis direction,
and the X and Y coordinates of the transformation origin.}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add text
}
\note{
\itemize{
\item \code{rsx_trans} for \code{add_text()} must have
the same number of rows as the total number of characters to be drawn;
not just the length of \code{text}.
\item Since textblobs do not have font fallback mechanism,
characters out of the specified font are not drawn correctly.
\item The return value is often a large object
because the specified font is embedded in the returned picture.
Note that you should almost always \code{\link[=freeze]{freeze()}} the picture after drawing text.
}
}
