% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-textpath.R
\name{add_textpath}
\alias{add_textpath}
\title{Add textpath}
\usage{
add_textpath(img, path, text, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{path}{A character vector of SVG notations
like \code{"M10 10 H 90 V 90 H 10 L 10 10"}.}

\item{text}{Characters to be drawn along \code{path}.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Draws text along SVG paths.

The return value is often a large object
because the specified font is embedded in the returned picture.
Note that you should almost always \code{\link[=freeze]{freeze()}} the picture after drawing text.
}
