% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-textpath.R
\name{add_textpath}
\alias{add_textpath}
\title{Add textpath}
\usage{
add_textpath(img, text, path, transform = diag(1, 3), ..., props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{text}{Characters to be drawn along \code{path}.}

\item{path}{A character vector of SVG notations
like \code{"M10 10 H 90 V 90 H 10 L 10 10"}.}

\item{transform}{Numerics of length 9
to apply affine transformations to the path themselves.
See \link{transform-matrix} for affine transformations.}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not take care of \code{...}, they are simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Draws text along SVG paths.

The return value is often a large object
because the specified font is embedded in the returned picture.
Note that you should almost always \code{\link[=freeze]{freeze()}} the picture after drawing text.
}
\details{
Since textblobs do not have font fallback mechanism,
characters out of the specified font are not drawn correctly.
}
