% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-vertices.R
\name{add_vertices}
\alias{add_vertices}
\title{Add vertices}
\usage{
add_vertices(img, vertices, ..., props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{vertices}{A double matrix where each row is a point.
If \code{nrow(vertices)} is not a multiple of 3,
the last \code{nrow(vertices) \%\% 3} points are ignored.}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add vertices
}
\details{
When \code{color} is explicitly provided, they will be combined with
\code{props[["shader"]]} if present or opaque \code{props[["color"]]} if not.

Therefore, if you want to paint each vertex with a different color,
you must specify a shader as \code{paint(shader = Shader$color(...))}.
}
