% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-vertices.R
\name{add_vertices}
\alias{add_vertices}
\title{Add vertices}
\usage{
add_vertices(img, vertices, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{vertices}{A double matrix where each row is a point.
If \code{nrow(vertices)} is not a multiple of 3,
the last \code{nrow(vertices) \%\% 3} points are ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
Add vertices
}
