% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{as_png}
\alias{as_png}
\title{Convert picture into PNG data}
\usage{
as_png(img, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{props}{A list of paint attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of PNG data.
}
\description{
Convert picture into PNG data
}
