% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{as_recordedplot}
\alias{as_recordedplot}
\title{Convert picture into recorded plot}
\usage{
as_recordedplot(img, ..., props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{...}{For some drawing functions, you can specify
\code{sigma}, \code{width} and \code{color} as named arguments.
\itemize{
\item \code{sigma} must be a numeric vector of blur sigmas for each shape.
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A \code{recordedplot} object is invisibly returned.
See \code{\link[grDevices:recordplot]{grDevices::recordPlot()}} for details.
}
\description{
This is mainly for testing purposes.
}
