% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skiagd-package.R
\name{as_recordedplot}
\alias{as_recordedplot}
\title{Convert a picture into a recorded plot}
\usage{
as_recordedplot(img, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}
}
\value{
A \code{recordedplot} object. See \code{\link[grDevices:recordplot]{grDevices::recordPlot()}} for details.
}
\description{
This is mainly for testing purposes.
}
