% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas}
\alias{canvas}
\title{Create new canvas}
\usage{
canvas(fill = "transparent", canvas_size = paint()[["canvas_size"]])
}
\arguments{
\item{fill}{RGBA representation of a color.
This can be specified using named colors or hexadecimal color codes,
which are converted internally using \code{\link[colorfast:col_to_rgb]{colorfast::col_to_rgb()}}.}

\item{canvas_size}{Integers of length 2.}
}
\value{
A raw vector of picture.
}
\description{
Creates a new canvas filled with specified color.
}
