% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas}
\alias{canvas}
\title{Create a canvas}
\usage{
canvas(fill = "transparent", size = paint()[["canvas_size"]])
}
\arguments{
\item{fill}{RGBA representation of a color.
This can be specified using named colors or hexadecimal color codes,
which are converted internally using \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}.}

\item{size}{Integers of length 2; canvas size.}
}
\value{
A raw vector of picture.
}
\description{
Creates a new canvas filled with specified color.
}
