% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{canvas}
\alias{canvas}
\title{Create a canvas}
\usage{
canvas(fill = paint()[["fill"]], .size = dev_size())
}
\arguments{
\item{fill}{A string; color name.}

\item{.size}{An integer vector.}
}
\value{
A raw vector.
}
\description{
Creates a new canvas filled with specified color.
}
