% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas}
\alias{canvas}
\title{Create a canvas}
\usage{
canvas(fill = paint()[["color"]], size = paint()[["canvas_size"]])
}
\arguments{
\item{fill}{A string; color name.}

\item{size}{An integer; canvas size.}
}
\value{
A raw vector of picture.
}
\description{
Creates a new canvas filled with specified color.
}
