% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas}
\alias{canvas}
\title{Create a canvas}
\usage{
canvas(fill = "transparent", size = paint()[["canvas_size"]])
}
\arguments{
\item{fill}{A string scalar; named colors or hexadecimal color codes.}

\item{size}{Integers of length 2; canvas size.}
}
\value{
A raw vector of picture.
}
\description{
Creates a new canvas filled with specified color.
}
