% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{col2rgba}
\alias{col2rgba}
\title{Convert colors to matrix of RGBA integers}
\usage{
col2rgba(color)
}
\arguments{
\item{color}{\code{col} for \code{\link[colorfast:col_to_rgb]{colorfast::col_to_rgb()}}.}
}
\value{
An integer matrix with 4 rows (RGBA) and N columns (the same length as \code{color}).
}
\description{
A wrapper of \code{\link[colorfast:col_to_rgb]{colorfast::col_to_rgb()}}.
}
