% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{col2rgba}
\alias{col2rgba}
\title{Color to RGBA}
\usage{
col2rgba(color)
}
\arguments{
\item{color}{\code{col} for \code{\link[colorfast:col_to_rgb]{colorfast::col_to_rgb()}}.}
}
\value{
An integer vector of length 4.
}
\description{
A wrapper of \code{\link[colorfast:col_to_rgb]{colorfast::col_to_rgb()}}.
}
