% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{create_mapping}
\alias{create_mapping}
\title{Create transformation matrix from points pairs}
\usage{
create_mapping(src, dst)
}
\arguments{
\item{src}{A numeric matrix (or a data-frame-like object)
with 2 columns (x and y) and 4 rows,
where each row is a point.}

\item{dst}{A numeric matrix (or a data-frame-like object)
with 2 columns (x and y) and 4 rows,
where each row is a point.}
}
\value{
A numeric matrix of size 3x3.
}
\description{
Creates a transformation matrix from sets of points \code{src} and \code{dst}.
If mapping is not possible, throws an error.
}
\examples{
\dontrun{
cv_size <- dev_size()

# Create an affine matrix that fits a 768x576 picture to current device size
create_mapping(
  src = matrix(
    c(0, 0, 768, 0, 768, 576, 0, 576),
  ),
  dst = matrix(
    c(0, 0, cv_size[1], 0, cv_size[1], cv_size[2], 0, cv_size[2]),
    ncol = 2,
    byrow = TRUE
  )
)
}
}
