% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{dev_size}
\alias{dev_size}
\title{Device size}
\usage{
dev_size(units = "px")
}
\arguments{
\item{units}{\code{units} for \code{\link[grDevices:dev.size]{grDevices::dev.size()}}.}
}
\value{
an integer vector.
}
\description{
Just returns the size of the current device as an integer (not a numeric).
}
