% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native-raster.R
\name{draw_img}
\alias{draw_img}
\title{Plot picture as a raster}
\usage{
draw_img(img, ..., props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{...}{For some drawing functions, you can specify
\code{sigma}, \code{width} and \code{color} as named arguments.
\itemize{
\item \code{sigma} must be a numeric vector of blur sigmas for each shape.
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
\code{img} is returned invisibly.
}
\description{
Plot picture as a raster
}
