% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{draw_img}
\alias{draw_img}
\title{Plot picture as a raster}
\usage{
draw_img(img, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.
For drawing functions, this can be a list of several painting attributes;
if it is a nested list, the drawing operation is performed
sequentially using each one.}
}
\value{
\code{img} is returned invisibly.
}
\description{
Plot picture as a raster
}
