% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{draw_img}
\alias{draw_img}
\title{Plot picture as PNG image}
\usage{
draw_img(img, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
\code{img} is returned invisibly.
}
\description{
Plot picture as PNG image
}
