% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{freeze}
\alias{freeze}
\title{Freeze a picture}
\usage{
freeze(img, props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{props}{A paint properties out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
\code{as_png(img)} and then adds it to a new transparent canvas.
}
