% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{freeze}
\alias{freeze}
\title{Freeze picture}
\usage{
freeze(img, left = 0, top = 0, fill = "transparent", props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{left}{Left offset for drawing PNG image.}

\item{top}{Top offset for drawing PNG image.}

\item{fill}{RGBA representation of a color.
This can be specified using named colors or hexadecimal color codes,
which are converted internally using \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
\code{as_png(img, props)} and then adds it to a new canvas
with the default blend mode (\code{BlendMode$SrcOver}).
}
