% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{freeze}
\alias{freeze}
\title{Freeze picture}
\usage{
freeze(img, fill = "transparent", props = paint())
}
\arguments{
\item{img}{A raw vector of picture.}

\item{fill}{A string scalar; named colors or hexadecimal color codes.}

\item{props}{A list of paint attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A raw vector of picture.
}
\description{
\code{as_png(img)} and then adds it to a new canvas.
}
