% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{paint}
\alias{paint}
\title{Define paint properties}
\usage{
paint(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
Named arguments specifying paint properties. See details.}
}
\value{
A list containing the specified paint properties,
merged with default values.
}
\description{
The \code{paint()} function allows users to specify
various properties for drawing shapes on the canvas,
such as color, stroke width, and transformations.
}
\details{
The following properties can be specified:
\itemize{
\item \code{canvas_size}: Integers of length 2 (width, height)
\item \code{color}: RGBA representation of a color. The color can be specified using named colors or hexadecimal color codes, which are converted internally using \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}.
\item \code{style}: \link{Style}
\item \code{join}: \link{Join}
\item \code{cap}: \link{Cap}
\item \code{width}: Numeric scalar (stroke width)
\item \code{miter}: Numeric scalar (stroke miter)
\item \code{blend_mode}: \link{BlendMode}
\item \code{path_effect}: \link{PathEffect}
\item \code{point_mode}: \link{PointMode} for \code{\link[=add_point]{add_point()}}
\item \code{transform}: Numerics of length 9. See \link{transform-matrix} for affine transformations.
}
}
