% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint.R
\name{paint}
\alias{paint}
\title{Define painting attributes}
\usage{
paint(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
Named arguments specifying painting attributes. See details.}
}
\value{
A list containing the specified painting attributes,
merged with default values.
}
\description{
The \code{paint()} function allows users to specify
various painting attributes for drawing shapes on the canvas,
such as color, stroke width, and transformations.
}
\details{
The following painting attributes can be specified:
\itemize{
\item \code{canvas_size}: Integers of length 2 (width, height).
\item \code{color}: RGBA representation of a color. The \code{color} can be specified using named colors or hexadecimal color codes, which are converted internally using \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}.
\item \code{style}: The paint style. See \link{Style}.
\item \code{join}: Stroke join. See \link{Join}.
\item \code{cap}: Stroke cap. See \link{Cap}.
\item \code{width}: A numeric scalar (stroke width).
\item \code{miter}: A numeric scalar (stroke miter).
\item \code{blend_mode}: See \link{BlendMode}.
\item \code{path_effect}: See \link{PathEffect}.
\item \code{shader}: See \link{Shader}.
\item \code{point_mode}: \link{PointMode} for \code{\link[=add_point]{add_point()}}.
\item \code{fill_type}: \link{FillType} for \code{\link[=add_path]{add_path()}}.
\item \code{transform}: Numerics of length 9. See \link{transform-matrix} for affine transformations.
}
}
