% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\name{save_img}
\alias{save_img}
\title{Save image to file}
\usage{
save_img(img, filename, .size = dev_size())
}
\arguments{
\item{img}{A raw vector.}

\item{filename}{A string; file name to save.}

\item{.size}{An integer vector.}
}
\value{
\code{filename} is invisibly returned.
}
\description{
Save image to file
}
