% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{Cap}
\alias{Cap}
\title{Cap (0-2)}
\format{
An object of class \code{Cap__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 3.
}
\usage{
Cap
}
\description{
\code{Cap} determines the stroke cap (the geometry drawn at the beginning and end of strokes).
}
\details{
The following caps are available:
\itemize{
\item \code{Butt}: Butt cap.
\item \code{Round}: Round cap.
\item \code{Square}: Square cap.
}
}
\seealso{
\href{https://rust-skia.github.io/doc/skia_safe/paint/enum.Cap.html}{Cap in skia_safe::paint - Rust}

Other paint-attributes: 
\code{\link{BlendMode}},
\code{\link{BlurStyle}},
\code{\link{FillType}},
\code{\link{FontStyle}},
\code{\link{ImageFilter}},
\code{\link{Join}},
\code{\link{PathEffect}},
\code{\link{PointMode}},
\code{\link{Shader}},
\code{\link{Style}},
\code{\link{VertexMode}}
}
\concept{paint-attributes}
\keyword{datasets}
