% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{FontStyle}
\alias{FontStyle}
\title{FontStyle (0-3)}
\format{
An object of class \code{FontStyle__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 4.
}
\usage{
FontStyle
}
\description{
\code{FontStyle} determines the font style.
}
\details{
The following styles are available:
\itemize{
\item \code{Normal}: Normal (plain).
\item \code{Bold}: Bold (bold).
\item \code{Italic}: Italic (italic).
\item \code{BoldItalic}: BoldItalic (bold.italic).
}
}
\seealso{
\href{https://rust-skia.github.io/doc/skia_safe/struct.FontStyle.html}{FontStyle in skia_safe - Rust}

Other paint-attributes: 
\code{\link{BlendMode}},
\code{\link{Cap}},
\code{\link{FillType}},
\code{\link{Join}},
\code{\link{PathEffect}},
\code{\link{PointMode}},
\code{\link{Shader}},
\code{\link{Style}}
}
\concept{paint-attributes}
\keyword{datasets}
