% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{TileMode}
\alias{TileMode}
\title{TileMode (0-3)}
\format{
An object of class \code{TileMode__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 4.
}
\usage{
TileMode
}
\description{
\code{TileMode} determines how the source is tiled for shaders.
This is not a paint attribute.
To specify \code{TileMode}, directly pass these pointers to shader functions.
}
\keyword{datasets}
