% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{TileMode}
\alias{TileMode}
\title{TileMode (0-3)}
\format{
An object of class \code{skiagd::TileMode__bundle} (inherits from \code{savvy_skiagd__sealed}) of length 4.
}
\usage{
TileMode
}
\description{
\code{TileMode} determines how the source is tiled for shaders.
This is not a paint attribute. To specify \code{TileMode}, directly pass these pointers to shader functions.
}
\details{
The following \code{TileMode} are available:
\itemize{
\item \code{Clamp}
\item \code{Repeat}
\item \code{Mirror}
\item \code{Decal}
}
}
\seealso{
\href{https://rust-skia.github.io/doc/skia_safe/enum.TileMode.html}{TileMode in skia_safe - Rust}
}
\keyword{datasets}
