% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skiagd-package.R
\docType{package}
\name{skiagd-package}
\alias{skiagd}
\alias{skiagd-package}
\title{skiagd: R Wrapper for 'rust-skia'}
\description{
A toy R wrapper for 'rust-skia' \url{https://github.com/rust-skia/rust-skia} (the Rust crate 'skia_safe' \url{https://rust-skia.github.io/doc/skia_safe/}, a binding for 'Skia' \url{https://skia.org/}).
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Other contributors:
\itemize{
  \item Jeroen Ooms [copyright holder]
}

}
\keyword{internal}
