% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{param-img-and-props}
\alias{param-img-and-props}
\title{Params for drawing functions}
\arguments{
\item{img}{A raw vector of picture.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}

\item{...}{For some drawing functions, you can specify \code{width} and \code{color} as named arguments.
\itemize{
\item \code{width} must be a numeric vector of stroke widths for each shape.
\item \code{color} must be an integer matrix where each \emph{\strong{column}} is an RGBA color for each shape.
}

If they are not provided as named arguments, they will be taken from \code{props}.
If the function does not matter them,  \code{...} is simply ignored.}
}
\description{
Params for drawing functions
}
\keyword{internal}
