% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{param-dot-size}
\alias{param-dot-size}
\alias{param-img-and-props}
\title{.size}
\arguments{
\item{.size}{An integer vector.}

\item{img}{A raw vector.}

\item{props}{A paint properties out of \code{\link[=paint]{paint()}}.}
}
\description{
.size

img and props
}
