% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{param-img-and-props}
\alias{param-img-and-props}
\title{img and props}
\arguments{
\item{img}{A raw vector of picture.}

\item{props}{A paint properties out of \code{\link[=paint]{paint()}}.}
}
\description{
img and props
}
\keyword{internal}
