% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shader.R
\name{Shader}
\alias{Shader}
\title{Shader}
\arguments{
\item{img}{A raw vector of picture.}

\item{mode}{For \code{blend()}, \link{BlendMode}. For others, \link{TileMode}.}

\item{tile_size}{Numerics of length 2; tile size (width, height).}

\item{transform}{Numerics of length 9; transformation matrix.}

\item{png}{A raw vector of PNG image.}

\item{rgba}{Integers of length 4 in range \verb{[0, 255]}
representing RGBA.}

\item{dst}{A \code{Shader} object; destination shader.}

\item{src}{A \code{Shader} object; source shader.}

\item{freq}{Numerics of length 2; frequencies.}

\item{octaves}{A numeric scalar; number of octaves.}

\item{seed}{Integer scalar; random seed.}

\item{start}{Numerics of length 2; starting point (x, y).}

\item{end}{Numerics of length 2; ending point (x, y).}

\item{from}{Integers of lenth 4 in range \verb{[0, 255]}; starting color.}

\item{to}{Integers of length 4 in range \verb{[0, 255]}; ending color.}

\item{flags}{A logical scalar; typically, you can leave this as \code{FALSE}.
See \href{https://shopify.github.io/react-native-skia/docs/shaders/gradients/#common-properties}{here}
for details.}

\item{radii}{Numerics of length 2; radii of start and end circles.}

\item{center}{Numerics of length 2; center of the gradient.}

\item{start_angle}{A numeric scalar in range \verb{[0, 360]};
starting angle. For default, set \code{0}.}

\item{end_angle}{A numeric scalar in range \verb{[0, 360]};
ending angle. For default, set \code{360}.}
}
\value{
A \code{Shader} object.
}
\description{
\code{Shader} is a struct that offers a reference to \code{skia_safe::Shader}.
You can apply a shader to shapes via \code{\link[=paint]{paint()}}.

Concatenating shaders with \code{c()} is equivalent to blend them all
into a single shader using \code{Shader$blend()} with the default \code{BlendMode}.
You can pass \code{mode} explicitly for \code{c()} to change the blend mode.
}
\details{
The following shaders are available:
\itemize{
\item \code{no_shader()}: does not apply any shader. This is the default shader for \code{\link[=paint]{paint()}}.
\item \code{from_picture(img, mode, tile_size, transform)}: takes a picture and returns an image shader.
\item \code{from_png(png, mode, transform)}: takes a PNG image and returns an image shader.
\item \code{color(rgba)}: takess a color and returns a color shader.
\item \code{blend(mode, dst, src)}: returns a shader that combines the given shaders with \link{BlendMode}.
\item \code{fractal_noise(freq, octaves, seed, tile_size)}: fractal perlin noise shader.
\item \code{turbulence(freq, octaves, seed, tile_size)}: turbulence noise shader.
\item \code{linear_gradient(start, end, from, to, mode, flags, transform)}: linear gradient shader.
\item \code{radial_gradient(center, radius, from, to, mode, flags, transform)}: radial gradient shader.
\item \code{conical_gradient(start, end, radii, from, to, mode, flags, transform)}: conical gradient shader.
\item \code{sweep_gradient(center, start_angle, end_angle, from, to, mode, flags, transform)}: sweep gradient shader.
}
}
\seealso{
\itemize{
\item \href{https://shopify.github.io/react-native-skia/docs/shaders/images}{Image Shaders | React Native Skia}
\item \href{https://shopify.github.io/react-native-skia/docs/shaders/gradients}{Gradients | React Native Skia}
\item \href{https://shopify.github.io/react-native-skia/docs/shaders/perlin-noise}{Perlin Noise Shaders | React Native Skia}
\item \href{https://shopify.github.io/react-native-skia/docs/shaders/colors}{Blending and Colors | React Native Skia}
}

Other paint-attributes: 
\code{\link{BlendMode}},
\code{\link{Cap}},
\code{\link{FillType}},
\code{\link{FontStyle}},
\code{\link{Join}},
\code{\link{PathEffect}},
\code{\link{PointMode}},
\code{\link{Style}},
\code{\link{VertexMode}}
}
\concept{paint-attributes}
