% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-utils.R
\name{svg_bounds}
\alias{svg_bounds}
\title{Retrieve bounding boxes of SVG paths}
\usage{
svg_bounds(path)
}
\arguments{
\item{path}{Characters of SVG notations.}
}
\value{
A tibble.
}
\description{
Retrieve bounding boxes of SVG paths
}
\seealso{
Other path-utils: 
\code{\link{svg_interpolate}()},
\code{\link{svg_transform}()}
}
\concept{path-utils}
