% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-textblob.R
\name{text_info}
\alias{text_info}
\title{Get width and number of characters}
\usage{
text_info(text, props = paint())
}
\arguments{
\item{text}{Text strings.}

\item{props}{A list of painting attributes out of \code{\link[=paint]{paint()}}.}
}
\value{
A tibble containing \code{id}, \code{n_chars} and \code{width}.
}
\description{
Returns information about text strings
when they are drawn naturally as a textblob with \code{props}.
}
