% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skiagd-package.R
\name{with_group}
\alias{with_group}
\title{Paint group}
\usage{
with_group(props, code)
}
\arguments{
\item{props}{Paint properties.}

\item{code}{Code to evaluate.}
}
\description{
Evaluates \code{code} with \code{props}.
}
